Record LTTng snapshot when latency higher than a threshold
When a request takes longer than the specified threshold (in ms), use
LTTng to record a snapshot of all kernel events in /tmp.

For now the snapshots are recorded locally (in /tmp), so the
lttng-consumerd requests appear in the iosnoop output every time we
record a snapshot, but it could be sent on the network to a remote
lttng-relayd.

Signed-off-by: Julien Desfossez <jdesfossez@efficios.com>